Add model classes
The plan is for emails to be logged in a database using Entity Framework Core.

The design of the model is relatively simple to allow for quick insertions
while still keeping track of all the necessary information. While one could
de-duplicate addresses by using a separate table (with a unique column),
this is not done as it would slow down the throughput.