Rename BaseAPI project to Mailer.Core
When I started the project I wasn't sure what to name it, so I named
each segment after it's purpose. I've decided to go with "Mailer"
as a simple name and rename the library project to Mailer.Core.
The OData API will be Mailer.OData.

Tests will remain tests for now.