Add IEmailTransport interface
The IEmailTransport interface is meant to abstract away the transporter
of an email message. This will allow Mailer.Core to be used with a
library other than MailKit if one wanted.

Since the Mailer example project needs the ability to retry sending
an email, a unified exception was needed for when retrying may
work (it would be a waste to retry if the failure occurred due to
missing credentials). TransportConnectionException exists for this
reason.