Add email services to Mailer.OData project
This adds the necessary services for sending an email by SMTP. I decided
it would best to ensure that sensitive data logging is disabled in
non-debug builds since that setting could be set by appsettings.json.

I have tested this by manually adding an email to the queue and it is
working.