Add NewEmailMessage to contain data in a new email
When a user sends a new email to send, we need to only grab the
properties that a user can provide. The best way to do this is to make
a separate type that excludes any property we don't want a user
setting.

In this case, we don't want a user setting WasSent (CreatedOn and
LastModifiedOn would be overwritten in EmailDbContext anyway).
Thankfully, this case is simple enough that we can create a separate
type and have EmailMessage inherit it.