Use database migration in Program.cs
For a while, the OData project has been manually handling database
creation for the sake of testing. Now that this example is almost
finished, this can be removed. The database is automatically migrated
in a development environment, but not any other. This is because a
production database should be manually migrated.